+++ file-size-baseline.json
@@ -15,7 +15,7 @@
     "open-sse/executors/muse-spark-web.ts": 1284,
     "open-sse/executors/perplexity-web.ts": 868,
     "open-sse/handlers/audioSpeech.ts": 952,
-    "open-sse/handlers/chatCore.ts": 6023,
+    "open-sse/handlers/chatCore.ts": 5808,
     "open-sse/handlers/imageGeneration.ts": 3777,
     "open-sse/handlers/responseSanitizer.ts": 1103,
     "open-sse/handlers/search.ts": 1442,
@@ -35,7 +35,7 @@
     "open-sse/translator/response/openai-responses.ts": 873,
     "open-sse/utils/cursorAgentProtobuf.ts": 1499,
     "open-sse/utils/stream.ts": 2710,
-    "src/app/(dashboard)/dashboard/HomePageClient.tsx": 1417,
+    "src/app/(dashboard)/dashboard/HomePageClient.tsx": 1385,
     "src/app/(dashboard)/dashboard/analytics/ComboHealthTab.tsx": 1020,
     "src/app/(dashboard)/dashboard/api-manager/ApiManagerPageClient.tsx": 2680,
     "src/app/(dashboard)/dashboard/cache/media/MediaPageClient.tsx": 1105,
@@ -48,7 +48,7 @@
     "src/app/(dashboard)/dashboard/endpoint/EndpointPageClient.tsx": 2570,
     "src/app/(dashboard)/dashboard/health/page.tsx": 1091,
     "src/app/(dashboard)/dashboard/playground/components/tabs/ApiTab.tsx": 847,
-    "src/app/(dashboard)/dashboard/providers/[id]/ProviderDetailPageClient.tsx": 3409,
+    "src/app/(dashboard)/dashboard/providers/[id]/ProviderDetailPageClient.tsx": 2554,
     "src/app/(dashboard)/dashboard/providers/[id]/components/ConnectionRow.tsx": 941,
     "src/app/(dashboard)/dashboard/providers/[id]/components/modals/AddApiKeyModal.tsx": 843,
     "src/app/(dashboard)/dashboard/providers/[id]/components/modals/EditConnectionModal.tsx": 1171,
@@ -73,7 +73,7 @@
     "src/app/api/oauth/[provider]/[action]/route.ts": 897,
     "src/app/api/providers/[id]/models/route.ts": 2426,
     "src/app/api/providers/[id]/test/route.ts": 842,
-    "src/app/api/usage/analytics/route.ts": 1355,
+    "src/app/api/usage/analytics/route.ts": 941,
     "src/app/api/v1/models/catalog.ts": 1435,
     "src/lib/cloudflaredTunnel.ts": 934,
     "src/lib/db/apiKeys.ts": 1490,
